% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtlasRangeEncode.R
\name{getAtlasRangeEncode}
\alias{getAtlasRangeEncode}
\title{query GRange from mongodb atlas}
\usage{
getAtlasRangeEncode(mychr, mystart, myend, url, mycoll, mydb)
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{url}{character string with connection url to MongoDB Atlas cluster}

\item{mycoll}{collection name in mongodb}

\item{mydb}{database name in mongodb}
}
\description{
query GRange from mongodb atlas
}
\examples{
\dontrun{
require(GenomicRanges)
require(mongolite)
mychrGo="chr11"
mystartGo=41196312-500*1000
myendGo=41322262+500*1000
mycollGo = "encode690"
mydbGo="encode"
url="mongodb+srv://test:test123!@cluster1-ag7nd.mongodb.net/test"
res=getAtlasRangeEncode(mychrGo,mystartGo,myendGo,mycollGo,mydbGo)
res
}
}
