% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtlasRangeEqtl.R
\name{getAtlasRangeEqtl}
\alias{getAtlasRangeEqtl}
\title{query eQTL from mongodb atlas}
\usage{
getAtlasRangeEqtl(mychr, mystart, myend, mycoll, mydb, url)
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{mycoll}{collection name in mongodb}

\item{mydb}{database name in mongodb}

\item{url}{character string with connection url to MongoDB Atlas cluster}
}
\description{
query eQTL from mongodb atlas
}
\examples{
\dontrun{
require(GenomicRanges)
require(mongolite)
mychrGo=17
mystartGo=41196312-500*1000
myendGo=41322262+500*1000
mycollGo = "Whole_Blood_allpairs_v7_eQTL"
mydbGo="txregnet"
url="mongodb+srv://test:test123!@cluster1-ag7nd.mongodb.net/test"
res_eqtl=getAtlasRangeEqtl(mychrGo,mystartGo,myendGo,mycollGo,mydbGo)
res_eqtl
}
}
