% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAtlasRangeFp.R
\name{getAtlasRangeFp}
\alias{getAtlasRangeFp}
\title{query Footprint data from mongodb atlas}
\usage{
getAtlasRangeFp(mychr, mystart, myend, url, mycoll, mydb)
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{url}{character string with connection url to MongoDB Atlas cluster}

\item{mycoll}{collection name in mongodb}

\item{mydb}{database name in mongodb}
}
\description{
query Footprint data from mongodb atlas
}
\examples{
\dontrun{
mychrGo="chr17"
mystartGo=41196312-500*1000
myendGo=41322262+500*1000
mycollGo="fSkin_fibro_bicep_R_DS19745_hg19_FP"
mydbGo="txregnet"
url="mongodb+srv://test:test123!@cluster1-ag7nd.mongodb.net/test"
res_fp=getAtlasRangeFp(mychrGo,mystartGo,myendGo,mycollGo,mydbGo)
res_fp
}
}
