% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrillTF.R
\name{getDrillTF}
\alias{getDrillTF}
\title{query TF data from Drill}
\usage{
getDrillTF(mychr, mystart, myend, myfile, myip, myport = 8047)
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{myfile}{path to TF file}

\item{myip}{system ip}

\item{myport}{port number}
}
\description{
query TF data from Drill
}
