% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongoRangeEqtl.R
\name{getMongoRangeEqtl}
\alias{getMongoRangeEqtl}
\title{query eQTL from mongodb}
\usage{
getMongoRangeEqtl(mychr, mystart, myend, mycoll, mydb)
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{mycoll}{collection name in mongodb}

\item{mydb}{database name in mongodb}
}
\description{
query eQTL from mongodb
}
\examples{
\dontrun{
require(GenomicRanges)
require(mongolite)
mychrGo=17
mystartGo=41196312-500*1000
myendGo=41322262+500*1000
mycollGo = "Breast_Mammary_Tissue_Analysis.v6p.all_snpgene_pairs_eQTL"
mydbGo="txregnet"
res_eqtl=getMongoRangeEqtl(mychrGo,mystartGo,myendGo,mycollGo,mydbGo)
res_eqtl
}
}
