% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongoRangeFp.R
\name{getMongoRangeFp}
\alias{getMongoRangeFp}
\title{query Footprint data from mongodb}
\usage{
getMongoRangeFp(mychr, mystart, myend, mycoll, mydb)
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{mycoll}{collection name in mongodb}

\item{mydb}{database name in mongodb}
}
\description{
query Footprint data from mongodb
}
\examples{
\dontrun{
mychrGo="chr17"
mystartGo=41196312-500*1000
myendGo=41322262+500*1000
mycollGo="vHMEC_DS18406_footprint"
mydbGo="txregnet"
res_fp=getMongoRangeFp(mychrGo,mystartGo,myendGo,mycollGo,myendGo)
res_fp
}
}
