% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoSource.R
\name{mongoSource}
\alias{mongoSource}
\title{connect to mongo database and retrieve collection names}
\usage{
mongoSource(db = "txregnet")
}
\arguments{
\item{db}{default database is txregnet}
}
\description{
connect to mongo database and retrieve collection names
}
