% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongoRangeEqtl.R
\name{getMongoRangeEqtl}
\alias{getMongoRangeEqtl}
\title{query eQTL from mongodb}
\usage{
getMongoRangeEqtl(mychr, mystart, myend,
  mycoll = "Breast_Mammary_Tissue_Analysis.v6p.all_snpgene_pairs_eQTL",
  mydb = "txregnet")
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{mycoll}{collection name in mongodb}

\item{mydb}{database name in mongodb}
}
\description{
query eQTL from mongodb
}
\examples{
\dontrun{
mychrGo=17
mystart=41196312-500*1000
myend=41322262+500*1000
res_eqtl=getMongoRangeEqtl(mychrGO,mystartGo,myendGo)
res_eqtl
}
}
