% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMongoRangeFp.R
\name{getMongoRangeFp}
\alias{getMongoRangeFp}
\title{query Footprint data from mongodb}
\usage{
getMongoRangeFp(mychr, mystart, myend, mycoll = "vHMEC_DS18406_footprint",
  mydb = "txregnet")
}
\arguments{
\item{mychr}{chromosome}

\item{mystart}{start position}

\item{myend}{end position}

\item{mycoll}{collection name in mongodb}

\item{mydb}{database name in mongodb}
}
\description{
query Footprint data from mongodb
}
\examples{
\dontrun{
mychrGo=17
mystart=41196312-500*1000
myend=41322262+500*1000
res_fp=getMongoRangeFp(mychrGO,mystartGo,myendGo)
res_fp
}
}
