% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongoSource.R
\name{mongoSource}
\alias{mongoSource}
\title{connect to mongo database and retrieve collection names}
\usage{
mongoSource(dbName, host, port)
}
\arguments{
\item{dbName}{character with database name}

\item{host}{character variable with url}

\item{port}{port number}
}
\description{
connect to mongo database and retrieve collection names
}
